+++ src/App.js
@@ -1,11 +1,11 @@
 import JunteSecao from './componentes/JunteSecao';
-import Carrosel from './componentes/Carrosel';
+import ProjetoCarrosel from './componentes/ProjetoCarrosel';
 
 function App() {
   return (
     <div className="App">   
       <JunteSecao /> 
-      <Carrosel />  
+      <ProjetoCarrosel />  
     </div>
   );
 }
