+++ src/componentes/Projeto/projeto.css
@@ -1,24 +1,3 @@
-.carrosel {
-    padding: 50px 150px;
-}
-
-.carrosel-titulo {
-    font-family: 'Poppins';
-    font-style: normal;
-    font-weight: 900;
-    font-size: 46px;
-    line-height: 69px;
-    color: #FE6C05;
-    text-align: center;
-    margin-bottom: 80px;
-}
-
-.projetos {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
 .projeto {
     max-width: 520px;
     background: #FBFBFB;
