+++ src/componentes/ProjetoCarrosel/projeto-carrosel.css
@@ -1,0 +1,20 @@
+.carrosel {
+    padding: 50px 150px;
+}
+
+.carrosel-titulo {
+    font-family: 'Poppins';
+    font-style: normal;
+    font-weight: 900;
+    font-size: 46px;
+    line-height: 69px;
+    color: #FE6C05;
+    text-align: center;
+    margin-bottom: 80px;
+}
+
+.projetos {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
